+++ logiciels/mikisa/poseidon/fluide.hh
@@ -45,7 +45,7 @@ inline auto est_fluide(Grille<int> const &flags, long idx)
 	return (flags.valeur(idx) & TypeFluid) != 0;
 }
 
-inline auto est_fluide(Grille<int> const &flags, size_t i, size_t j, size_t k)
+inline auto est_fluide(Grille<int> const &flags, long i, long j, long k)
 {
 	return est_fluide(flags, flags.calcul_index(i, j, k));
 }
@@ -55,7 +55,7 @@ inline auto est_obstacle(Grille<int> const &flags, long idx)
 	return (flags.valeur(idx) & TypeObstacle) != 0;
 }
 
-inline auto est_obstacle(Grille<int> const &flags, size_t i, size_t j, size_t k)
+inline auto est_obstacle(Grille<int> const &flags, long i, long j, long k)
 {
 	return est_obstacle(flags, flags.calcul_index(i, j, k));
 }
@@ -65,7 +65,7 @@ inline auto est_vide(Grille<int> const &flags, long idx)
 	return (flags.valeur(idx) & TypeVide) != 0;
 }
 
-inline auto est_vide(Grille<int> const &flags, size_t i, size_t j, size_t k)
+inline auto est_vide(Grille<int> const &flags, long i, long j, long k)
 {
 	return est_vide(flags, flags.calcul_index(i, j, k));
 }
@@ -75,7 +75,7 @@ inline auto est_outflow(Grille<int> const &flags, long idx)
 	return (flags.valeur(idx) & TypeOutflow) != 0;
 }
 
-inline auto est_outflow(Grille<int> const &flags, size_t i, size_t j, size_t k)
+inline auto est_outflow(Grille<int> const &flags, long i, long j, long k)
 {
 	return est_outflow(flags, flags.calcul_index(i, j, k));
 }
